Reorganize a bit with 1 file/type & 1 mod/file
This little reorganization seems to be a good idea to keep anything
a bit clearer and maybe easier to use.
