+++ src/operators.rs
@@ -1,5 +1,3 @@
-use crate::*;
-
 #[derive(Debug)]
 pub enum Operators
 {
