+++ src/token.rs
@@ -1,0 +1,8 @@
+use crate::{fact::Fact, operation::Operation};
+
+#[derive(Debug)]
+pub enum Factoken
+{
+    Fact(Fact),
+    Operation(Operation),
+}
